refactor(tauri): split shell.rs into shell/ submodule (Phase 5)
Carves the 1,328-line `src-tauri/src/shell.rs` into three cohesive
files plus a thin `mod.rs`:

- `shell/sharemode.rs` — `ShareMode` enum + `ShareState` privacy-floor
  logic. Pure data + transitions (3 tests).
- `shell/scrollback.rs` — `Scrollback` ring buffer with monotonic
  cursor + forward-paging `read_from` (8 tests).
- `shell/lifecycle.rs` — `ShellRegistry`, `ShellSlot`, `ShellOpenArgs`,
  the four PTY commands + the four share-mode/scrollback commands,
  reader thread, `parse_osc_title`, `utf8_safe_split` (24 tests).
- `shell/mod.rs` — `pub use lifecycle::*` so
  `tauri::generate_handler![shell::shell_open, …]` resolves both the
  command function and the macro-generated `__cmd__*` /
  `__tauri_command_name_*` siblings.

No behavior change. All 73 lib tests pass; `cargo clippy -D warnings`,
`tsc -b`, ESLint, and 305-test vitest suite are clean.

Refs: #42, plan Phase 5.
